duplicate_keyspace.py
This python script is used to duplicate/copy source keyspace to target keyspace (it creates the tables and also copies the data into the new keyspace). The script accepts 3 inputs - name of the cluster, source and target keyspaces.

This script has been tested on local server with cluster spun up with ccm. The path of the datafiles is hardcoded, this will need to be changed. Also there has not be thorough validations put in the script.

This needs to be thoroughly tested and yet to be used in real environments.